HybridWebViewPage.xaml.CS: RegisterAction now uses ExecuteFromBrowser(data) function to receive commands from the JS Bridge. ExecuteFromBrowser() --- when we get a "quit" message, use MessagingCenter to send "QuitApp" message. Implement OnBackButtonPressed() override to intercept hardware back button on Android, send "GoBack" message, and prevent default behavior.
DROID:

Droid-> Resources -> Resource.designer.cs: It looks like some color values got changed , but it wasn't me.

Added BackButtonAttemptCallback.cs as a Javascript IValueCallback to get the result of WiseJ Webmethod. It fire but the result is always null.

HybridWebViewRenderer.cs:
Subscribe to "GoBack" messages in OnElementChanges (when Control==null), and use webView.EvaluateJavascript to call the WiseJ webmethod "HitFormBackButtonIfPossible".

Subscribe to "QuitApp" message and halt the app when received.